Contact Form
(Re)aded contact form.  Had a contact form in before but it was super messed up and I couldn't be asked to fix it, so I just deleted it.  Then, decided it is a good idea and built it from the ground up in a different way.  Still no idea how functional it will be, but I'm guessing it will be hella functional, so that's fun.