Tested Contact Form
I tested the form finally to avoid continuing to look like a total wank by running it in a node.JS server to at least verify the JavaScript works like I want it to.  Still don't think this will work because the ole netlify seems wanky, but we shall see I guess.