Convert to Jekyll Blog!
I've been testing and modifying this for a little bit now, and I feel it's at a good place to toss it up onto my actual website now!  I am going to do a lot more editing of themes and content still, but it's at a comfortable place to make it live now.