Image Changes
Added my own images to replace the default images for the theme I used to get me started with this.  In case you were wondering, all of the photos were taken by my hand, edited with lighroom, and are of the gorgeous NC mountains I am so lucky to call home.